emoji.adminList APIへの問合せ処理を fetchEmojiAdminList.js モジュールに切り出した
